Example junctions

Junction co-ordinates and counts associated with the SOD1 gene. Junctions counts originate from GTEx liver samples and are downloaded via the Bioconductor package snapcount. Only unannotated junctions with a mean count above 0.3 have been retained for this example.

Usage

sod1_junctions

Format

A tibble::tibble():

seqnames

factor() chromosome.

start

integer() start position.

end

integer() end position.

strand

factor() strand.

mean_count

factor() Average count across all GTEx liver samples.

Source

generated using ggtranscript/data-raw/sod1_junctions.R