Convert exon co-ordinates to introns

Taking a set of exons, to_intron will return the corresponding introns. to_intron expects inputted exons to either 1. originate from a single transcript or 2. be grouped via group_var such that each group corresponds to a single transcript. In other words, to_intron assumes that exons (from each group) do not overlap one another.

to_intron(x, group_var = NULL, start_var = start, end_var = end)

Arguments

x

data.frame containing exons co-ordinates (and possibly associated transcript)

group_var

variable name (not in quotes) specifying the column with the group, most often the transcript id/name.

start_var

variable name (not in quotes) specifying the column with the start co-ordinate(s).

end_var

variable name (not in quotes) specifying the column with the end co-ordinate(s).

Examples

example_exons <-
    dplyr::tibble(
        start = c(5, 10, 15, 20),
        end = c(7, 12, 17, 22),
        tx = c("A", "A", "B", "B")
    )

example_exons
#> # A tibble: 4 × 3
#>   start   end tx   
#>   <dbl> <dbl> <chr>
#> 1     5     7 A    
#> 2    10    12 A    
#> 3    15    17 B    
#> 4    20    22 B    

to_intron(example_exons, group_var = tx)
#> # A tibble: 2 × 3
#>   tx    intron_start intron_end
#>   <chr>        <dbl>      <dbl>
#> 1 A                8          9
#> 2 B               18         19