ggtranscript: Visualizing transcript structure and annotation using ggplot2

The goal of ggtranscript is the simplify the process of visualizing transcript structure and annotation. To achieve this, ggtranscript introduces 5 new geoms (geom_range(), geom_half_range(), geom_intron(), geom_junction() and geom_junction_label_repel()) as well as several helper functions. As a ggplot2 extension, ggtranscript inherits ggplot2's familiarity and flexibility, enabling users to intuitively adjust aesthetics, parameters, scales etc as well as complement ggtranscript geoms with existing ggplot2 geoms to create informative, publication-ready plots.